Various edits to existing files

+++ commands/emote.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-unused-vars */
 exports.run = async (client, message, args, level) => {
+    if (!args[0]) return message.channel.send("You need to specify an emote to use!");
     const search = args[0].toLowerCase();
-    const emote = client.emojis.find(emote => emote.name === search);
+    const emote = client.emojis.find(emote => emote.name.toLowerCase().includes(search));
     if (!emote) return message.channel.send("That's not a valid emote name!");
     message.delete();
     message.channel.send(`${emote}`);
